Multiple
Made a general implementation of the levelstate. Fixed a wrong init call
on the levelstate, should fix the problem of getting a double log of the
player creation(it was being created twice) and fixed an unsafe thread
implementation causing crashes. Added menu button for multiplayer mode.
Removed obsolete states
<bug, feature, organisation, performance